+++ igzip/flatten_ll.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-void flatten_ll(uint16_t *ll_hist);
+void flatten_ll(uint32_t *ll_hist);
